Week 6/7 Homework_fall forces
Since my last week sketch was broken, I was not able to find the bug I decided to attempt something else.

The concept behind the sketch:
This is my first winter in NYC, from where im from we dont get to see the change in weather, so i made a sketch showing how fall is turning to a gloomy winter afternoon.

Problems:
When the particles draw theres a blue hue that covers the entire canvas, its the shade of the particles that signify snow in the wind but Im not sure why the colour is covering the entire canvas and not just the particles.